agent-script/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.50.0
google-cloud-storage>=2.0.0